implement multi-gpu training in HPO with Ray Tune and Ray Train
